Fixed branch current probing in instances with $mfactor!=1
According to LRM I(branch) should return the branch current divided by $mfactor.
This means that it should return the branch current of a single parallel instance.

Before this fix OpenVAF was dividing the value of the unknown by $mfactor.
Because a flow unknown in the system of equations represents the branch current
of a single parallel instance, I(branch) was returning the branch current
divided by squared $mfactor, i.e. the branch current of a single instance
divided by $mfactor. This was wrong.