Instance parameter defaulting fixed.
If an instance parameter was given with a model
defaulting did not work correctly. The default
value of an instance parameter is now stored
in the model data structure (if not given).
In instance data structure the default value is
also stored if parameter is not given. The default
value is taken from the model (if given) or from
parameter definition.